add a note to the dynamic form request panel
THe form request panel is generated on-the-fly, so we need a MutationObserver to detect when the element is created, and append the note when that happens.  Bad practice to modify the DOM in an Angular controller, but we don't have access to the parent control to modify it directly in the view model.